Rework diagram design
Since React does not support  inheritance I changed our previous design
of how diagrams work. Instead of inheritance I use composition.
Now the Diagram class is just responsible for the view of the diagram.
All the logic is outsourced to a DiagramController. IDiagramController
is a inteface which can be implemented by different diagram controllers
to provide different types of diagrams.