Improve code quality of dControllers and fix bugs
In the previous version the current configuration of a diagram was
stored in the view of a diagram. As this is not compatible with the MVC
architecture this job is now done by the controller itself.

A bug occured if a observationstation providet too many observations per
day. Hence a diagram only requested observations once per day, some
observations got missing in the overflow since the limit of 1000
observations was exceeded.
Now a diagram requests observations 4 times a day so that all
values make it into the diagram.

In addition to this changes I removed the
YearComparisonLineChartController we will no longer support
that type of diagram.

I also renamed the ComparisonToLastYearPieChartController
to ComparisonToLastMonthPieChartController since the timespan has
changed.